Move Request Insight key helpers to a neutral shared/lib module
